sv-server/requirements.txt
fastapi
uvicorn
python-multipart
python-slugify
